+++ src/ZplCommand/PrintStart.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Janisvepris\ZplBuilder\ZplCommand;
+
+use Janisvepris\ZplBuilder\ZplCommand;
+
+readonly class PrintStart implements ZplCommand
+{
+    public const string COMMAND = '~PS';
+    public const string FORMAT = '';
+
+    public function __toString()
+    {
+        return self::COMMAND;
+    }
+}